Frontend: Restructure frontend parser tests by node and scope
