+++ loki/frontend/tests/test_scope_associate.py
@@ -1,0 +1,179 @@
+# (C) Copyright 2018- ECMWF.
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+"""
+Verify correct frontend behaviour for associate-related scope handling.
+"""
+
+import numpy as np
+import pytest
+
+from loki import Module, Subroutine, BasicType
+from loki.jit_build import jit_compile
+from loki.expression import symbols as sym
+from loki.frontend import available_frontends, OMNI
+from loki.ir import nodes as ir, FindNodes, FindVariables
+
+
+@pytest.mark.parametrize('frontend', available_frontends())
+def test_associates(tmp_path, frontend):
+    """Test the use of associate to access and modify other items"""
+
+    fcode = """
+module derived_types_mod
+  integer, parameter :: jprb = selected_real_kind(13,300)
+
+  type explicit
+    real(kind=jprb) :: scalar, vector(3), matrix(3, 3)
+    real(kind=jprb) :: red_herring
+  end type explicit
+
+  type deferred
+    real(kind=jprb), allocatable :: scalar, vector(:), matrix(:, :)
+    real(kind=jprb), allocatable :: red_herring
+  end type deferred
+contains
+
+  subroutine alloc_deferred(item)
+    type(deferred), intent(inout) :: item
+    allocate(item%vector(3))
+    allocate(item%matrix(3, 3))
+  end subroutine alloc_deferred
+
+  subroutine free_deferred(item)
+    type(deferred), intent(inout) :: item
+    deallocate(item%vector)
+    deallocate(item%matrix)
+  end subroutine free_deferred
+
+  subroutine associates(item)
+    type(explicit), intent(inout) :: item
+    type(deferred) :: item2
+
+    item%scalar = 17.0
+
+    associate(vector2=>item%matrix(:,1))
+        vector2(:) = 3.
+        item%matrix(:,3) = vector2(:)
+    end associate
+
+    associate(vector=>item%vector)
+        item%vector(2) = vector(1)
+        vector(3) = item%vector(1) + vector(2)
+        vector(1) = 1.
+    end associate
+
+    call alloc_deferred(item2)
+
+    associate(vec=>item2%vector(2))
+        vec = 1.
+    end associate
+
+    call free_deferred(item2)
+  end subroutine associates
+end module
+"""
+    # Test the internals
+    module = Module.from_source(fcode, frontend=frontend, xmods=[tmp_path])
+    routine = module['associates']
+    variables = FindVariables().visit(routine.body)
+    assert all(
+        v.shape == ('3',) for v in variables if v.name in ['vector', 'vector2']
+    )
+
+    for assoc in FindNodes(ir.Associate).visit(routine.body):
+        for var in FindVariables().visit(assoc.body):
+            if var.name in assoc.variables:
+                assert var.scope is assoc
+                assert var.type.parent is None
+            else:
+                assert var.scope is routine
+
+    # Test the generated module
+    filepath = tmp_path/(f'derived_types_associates_{frontend}.f90')
+    mod = jit_compile(module, filepath=filepath, objname='derived_types_mod')
+
+    item = mod.explicit()
+    item.scalar = 0.
+    item.vector[0] = 5.
+    item.vector[1:2] = 0.
+    mod.associates(item)
+    assert item.scalar == 17.0 and (item.vector == [1., 5., 10.]).all()
+
+
+@pytest.mark.parametrize('frontend', available_frontends(xfail=[(OMNI, 'OMNI fails to read without full module')]))
+def test_associates_deferred(frontend):
+    """
+    Verify that reading in subroutines with deferred external type definitions
+    and associates working on that are supported.
+    """
+
+    fcode = """
+SUBROUTINE ASSOCIATES_DEFERRED(ITEM, IDX)
+USE SOME_MOD, ONLY: SOME_TYPE
+IMPLICIT NONE
+TYPE(SOME_TYPE), INTENT(IN) :: ITEM
+INTEGER, INTENT(IN) :: IDX
+ASSOCIATE(SOME_VAR=>ITEM%SOME_VAR(IDX), SOME_OTHER_VAR=>ITEM%SOME_VAR(ITEM%OFFSET))
+SOME_VAR = 5
+END ASSOCIATE
+END SUBROUTINE
+    """
+    routine = Subroutine.from_source(fcode, frontend=frontend)
+    variables = {v.name: v for v in FindVariables().visit(routine.body)}
+    assert len(variables) == 6
+    some_var = variables['SOME_VAR']
+    assert isinstance(some_var, sym.DeferredTypeSymbol)
+    assert some_var.name.upper() == 'SOME_VAR'
+    assert some_var.type.dtype == BasicType.DEFERRED
+    associate = FindNodes(ir.Associate).visit(routine.body)[0]
+    assert some_var.scope is associate
+
+    some_other_var = variables['SOME_OTHER_VAR']
+    assert isinstance(some_var, sym.DeferredTypeSymbol)
+    assert some_other_var.name.upper() == 'SOME_OTHER_VAR'
+    assert some_other_var.type.dtype == BasicType.DEFERRED
+    assert some_other_var.type.shape == ('ITEM%OFFSET',)
+    assert some_other_var.scope is associate
+
+
+@pytest.mark.parametrize('frontend', available_frontends())
+def test_associates_expr(tmp_path, frontend):
+    """Verify that associates with expressions are supported"""
+    fcode = """
+subroutine associates_expr(in, out)
+  implicit none
+  integer, intent(in) :: in(3)
+  integer, intent(out) :: out(3)
+
+  out(:) = 0
+
+  associate(a=>1+3)
+    out(:) = out(:) + a
+  end associate
+
+  associate(b=>2*in(:) + in(:))
+    out(:) = out(:) + b(:)
+  end associate
+end subroutine associates_expr
+    """.strip()
+    routine = Subroutine.from_source(fcode, frontend=frontend)
+
+    variables = {v.name: v for v in FindVariables().visit(routine.body)}
+    assert len(variables) == 4
+    assert isinstance(variables['a'], sym.DeferredTypeSymbol)
+    assert variables['a'].type.dtype is BasicType.DEFERRED  # TODO: support type derivation for expressions
+    assert isinstance(variables['b'], sym.Array)  # Note: this is an array because we have a shape
+    assert variables['b'].type.dtype is BasicType.DEFERRED  # TODO: support type derivation for expressions
+    assert variables['b'].type.shape == ('3',)
+
+    filepath = tmp_path/(f'associates_expr_{frontend}.f90')
+    function = jit_compile(routine, filepath=filepath, objname=routine.name)
+    a = np.array([1, 2, 3], dtype='i')
+    b = np.zeros(3, dtype='i')
+    function(a, b)
+    assert np.all(b == [7, 10, 13])
